fixup! JAMES-2345 Ensure quota is displayed for every mailboxes
